fake_qiubai/requirements.txt
beautifulsoup4==4.5.1
Django==1.10.1
django-braces==1.9.0
django-cors-headers==1.2.2
django-filter==0.15.2
django-oauth-toolkit==0.10.0
djangorestframework==3.4.7
gunicorn==19.6.0
Markdown==2.6.7
MySQL-python==1.2.5
Pillow==3.4.2
Pygments==2.1.3
redis==2.10.5
